# docs/swagger.yaml
basePath: /
definitions:
...
  service.LogInRequest:
    properties:
      password:
        description: '"Password"'
        example: zxc123zxc123
        type: string
      user_id:
        description: '"User Id"'
        example: gooddy20
        type: string
    required:
    - password
    - user_id
    type: object
  service.LogInResponse:
    properties:
      IsLogIn:
        description: '"true" = Pass, "false" = Incorrect "User Id" or "Password"'
        example: true
        type: boolean
    type: object
...
  service.NewUserRequest:
    properties:
      carb:
        description: Default carb (g.) of the "User"
        example: 120
        type: number
      fat:
        description: Default fat (g.) of the "User"
        example: 60
        type: number
      password:
        description: '"Password"'
        example: zxc123zxc123
        type: string
      protein:
        description: Default protein (g.) of the "User"
        example: 120
        type: number
      user_id:
        description: '"User Id"'
        example: gooddy20
        type: string
      username:
        description: '"Username"'
        example: GoodDy
        type: string
      weight:
        description: Default weight (kg.) of the "User"
        example: 70
        type: number
    required:
    - password
    - user_id
    - username
    type: object
...
  service.UpdateUserRequest:
    properties:
      carb:
        description: Carb that you want to change to
        example: 160
        type: number
      fat:
        description: Fat (g.) that you want to change to
        example: 70
        type: number
      favorite_menues:
        description: Favorite Menues's id that you want to change to e.g. "9,10" 9
          = "Moo Yang" and 10 = "Sticky Rice" so this "User" got "Moo Yang" and "Sticky
          Rice" as "Favorite Menu"
        example: 4,7,9,10,11
        type: string
      password:
        description: '"Password" that you want to change'
        example: zxc123zxc456
        type: string
      protein:
        description: Protein (g.) that you want to change to
        example: 150
        type: number
      user_id:
        description: '"User Id"'
        example: gooddy20
        type: string
      username:
        description: '"Username" that you want to change to'
        example: GooDDy19
        type: string
      weight:
        description: Weight (kg.) that you want to change to
        example: 72
        type: number
    required:
    - user_id
    type: object
...
  service.userResponse:
    properties:
      carb:
        description: Default carb (g.) of the "User"
        example: 130
        type: number
      fat:
        description: Default fat (g.) of the "User"
        example: 40
        type: number
      favorite_menues:
        description: Favorite Menues's id e.g. "9,10" 9 = "Moo Yang" and 10 = "Sticky
          Rice" so this "User" got "Moo Yang" and "Sticky Rice" as "Favorite Menu"
        example: 9,10
        type: string
      protein:
        description: Default protein (g.) of the "User"
        example: 140
        type: number
      username:
        description: '"Username"'
        example: GoodDy
        type: string
      weight:
        description: Default weight (kg.) of the "User"
        example: 62
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for record all meal that you have in each day and help you calculate
    summary nutrition in each meal and you can save favorite menu and favorite meal
    for track your diet easily and create your own menu
  title: Nutrition Calculator API documentation
  version: 1.0.0
paths:
...
  /user/:
    post:
      consumes:
      - application/json
      description: Create a `User`
      parameters:
      - description: '`User`''s data detail'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.NewUserRequest'
      responses:
        "200":
          description: OK
        "406":
          description: Request Body Not Acceptable
        "500":
          description: Internal Server Error
      summary: Create a "User"
      tags:
      - User
  /user/{user_id}:
    get:
      description: Get a `User`'s detail by `User Id`
      parameters:
      - description: '`User Id`'
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.userResponse'
        "406":
          description: '`User Id` is not found'
        "500":
          description: Internal Server Error
      summary: Get a "User"'s detail
      tags:
      - User
  /user/login:
    put:
      consumes:
      - application/json
      description: Check `User Id` and `Password` are correct or not
      parameters:
      - description: '`User Id` and `Password`'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.LogInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LogInResponse'
        "406":
          description: Request Body Not Acceptable
        "500":
          description: Internal Server Error
      summary: Check "User Id" and "Password" are correct or not
      tags:
      - User
  /user/userdetail:
    put:
      description: Update a `User`'s detail
      parameters:
      - description: '`User`''s data detail that you want to update and can ignore
          the unchanged parameters'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserRequest'
      responses:
        "200":
          description: OK
        "406":
          description: Request Body Not Acceptable or `User Id` is not found
        "500":
          description: Internal Server Error
      summary: Update a "User"'s detail
      tags:
      - User
swagger: "2.0"